requirements.txt
requests==2.31.0
django-cors-headers==4.3.1
Django==5.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
setuptools>=65.0.0
python-decouple==3.8
pymongo==4.6.0
gunicorn==21.2.0
whitenoise==6.6.0
